+++ include/spectroscopy.h
@@ -9,8 +9,9 @@
 
 #include "Boundary.h"
 #include "Particles.h"
-#include "math.h"
+#include <cmath>
 #include <vector>
+
 #if USE_CUDA > 0
 #if __CUDA_ARCH__ < 600
 __device__ double atomicAdd1(double *address, double val) {
@@ -72,7 +73,7 @@ struct spec_bin {
     float dim1 = particlesPointer->xprevious[indx];
 #else
 #if USECYLSYMM > 0
-    float dim1 = sqrtf(x * x + y * y);
+    float dim1 = std::sqrt(x * x + y * y);
 #else
     float dim1 = x;
 #endif
@@ -83,16 +84,16 @@ struct spec_bin {
         dx = gridX[1] - gridX[0];
         dz = gridZ[1] - gridZ[0];
 #if SPECTROSCOPY < 3
-        int indx_X = floor((dim1 - gridX[0]) / dx);
-        int indx_Z = floor((z - gridZ[0]) / dz);
+        int indx_X = std::floor((dim1 - gridX[0]) / dx);
+        int indx_Z = std::floor((z - gridZ[0]) / dz);
         int indx_Y = 0;
         int nnYY = 1;
 #else
         if ((y > gridY[0]) && (y < gridY[nY - 1])) {
-          int indx_X = floor((dim1 - gridX[0]) / dx);
-          int indx_Z = floor((z - gridZ[0]) / dz);
+          int indx_X = std::floor((dim1 - gridX[0]) / dx);
+          int indx_Z = std::floor((z - gridZ[0]) / dz);
           dy = gridY[1] - gridY[0];
-          int indx_Y = floor((y - gridY[0]) / dy);
+          int indx_Y = std::floor((y - gridY[0]) / dy);
           if (indx_Y < 0 || indx_Y >= nY)
             indx_Y = 0;
           int nnYY = nY;
@@ -107,7 +108,7 @@ struct spec_bin {
         //std::cout << "dx " << dx << std::endl;
         //std::cout << "dz " << dz << std::endl;
         //std::cout << "ind x " << indx_X << "ind z " << indx_Z << std::endl;
-        int charge = floor(particlesPointer->charge[indx]);
+        int charge = std::floor(particlesPointer->charge[indx]);
         if (particlesPointer->hitWall[indx] == 0.0) {
           float specWeight = particlesPointer->weight[indx];
 #if USE_CUDA > 0
